altt

+++ PI_Start/PI_Start/ENUMwPrunning.h
@@ -11,7 +11,7 @@
 
 #include "simpleMath.h"
 
-extern int dim;
+extern short dim;
 extern double **mu;
 extern double *B;
